requirements.txt
GitPython
PyGithub
python-dotenv
pytest